Fix test cleanup

+++ ext/pdo_sqlite/tests/pdo_sqlite_open_flags.phpt
@@ -5,23 +5,23 @@ PDO_sqlite: Testing open flags
 --FILE--
 <?php
 
-$filename = tempnam(sys_get_temp_dir(), 'sqlite');
+$filename = dirname(__FILE__) . DIRECTORY_SEPARATOR . "pdo_sqlite_open_flags.db";
 
 // Default open flag is read-write|create
 $db = new PDO('sqlite:' . $filename, null, null, [PDO::ATTR_ERRMODE => PDO::ERRMODE_EXCEPTION]);
 
 var_dump($db->exec('CREATE TABLE test1 (id INT);'));
 
-try {
-	$db = new PDO('sqlite:' . $filename, null, null, [PDO::SQLITE_ATTR_OPEN_FLAGS => PDO::SQLITE_OPEN_READONLY, PDO::ATTR_ERRMODE => PDO::ERRMODE_EXCEPTION]);
+$db = new PDO('sqlite:' . $filename, null, null, [PDO::SQLITE_ATTR_OPEN_FLAGS => PDO::SQLITE_OPEN_READONLY, PDO::ATTR_ERRMODE => PDO::ERRMODE_EXCEPTION]);
 
-	var_dump($db->exec('CREATE TABLE test2 (id INT);'));
-}
-finally {
-	// Cleanup
+var_dump($db->exec('CREATE TABLE test2 (id INT);'));
+?>
+--CLEAN--
+<?php
+$filename = dirname(__FILE__) . DIRECTORY_SEPARATOR . "pdo_sqlite_open_flags.db";
+if (file_exists($filename)) {
 	unlink($filename);
 }
-
 ?>
 --EXPECTF--
 int(0)
